requirements.txt
pandas 
openpyxl
openai
PyGithub
chardet
dotenv